+++ CHANGES.md
@@ -38,6 +38,10 @@ Fixed Issues:
 * [#16845](https://dev.ckeditor.com/ticket/16845): [IE] Fixed: Cursor jumps to the top of the scrolled editor after focusing it when [Copy Formatting](http://ckeditor.com/addon/copyformatting) plugin is enabled.
 * [#16927](https://dev.ckeditor.com/ticket/16927): Fixed: CKEditor throws an error if a bundle containing [Color Button](http://ckeditor.com/addon/colorbutton) plugin is run in ES5 strict mode. Thanks to [Igor Rubinovich](https://github.com/IgorRubinovich)!
 
+Other Changes:
+
+* [#16954](http://dev.ckeditor.com/ticket/16954): Remove paste dialog.
+
 ## CKEditor 4.6.2
 
 New Features: